+++ docs/CHANGELOG.md
@@ -5,6 +5,7 @@
 #### Added
 - `invite` command.
 - Error when giveaway result is too big is handled.
+- `avatar`'s embed description simplified.
 #### Changed
 - Using `giveaway` command requires "manage guild" permission.
 - Increased `meme`'s cooldown from 4s to 6s.