+++ docs/CHANGELOG.md
@@ -19,6 +19,7 @@
 - Profile's now are on field instead of big description.
 - Optional 3rd argument in `giveaway` is now asked for.
 - Final `giveaway` massage is prettier.
+- Cooldowns' messages no more include time, because it was incorrect.
 #### Fixed
 - First `daily` is not creating profile **and** adds cash.
 - Value 0 is now possible in `manage` command.